fix favorite

+++ queries/favoriteQueries.js
@@ -1,12 +1,7 @@
-export const addFavoriteQuery = `BEGIN;
+export const addFavoriteQuery = `INSERT INTO favorites (user_id, car_id)`;
+export const deleteFavoriteQuery = `DELETE FROM favorites WHERE user_id = $1 AND car_id= $2`;
 
-WITH deleted AS (
-    DELETE FROM favorites
-    WHERE user_id = $1 AND car_id = $2
-    RETURNING *
-)
-INSERT INTO favorites (user_id, car_id)
-SELECT $1, $2
-WHERE NOT EXISTS (SELECT 1 FROM deleted);
-
-COMMIT`;
+module.exports = {
+  addFavoriteQuery,
+  deleteFavoriteQuery,
+};
